pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "forestfires_project"
version = "0.0.1"
description = "Project description for forestfire project"
authors = [
  { name = "MLOPS-Group34", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dvc>=3.66.1",
    "dvc-gdrive>=3.0.1",
    "fastapi==0.115.6",
    "ipykernel>=7.1.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "protobuf>=6.33.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "torch==2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "typer==0.15.1",
    "ultralytics>=8.3.251",
    "uvicorn==0.34.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "ultralytics>=8.3.251",
    "roboflow>=1.2.11",
    "ruamel-yaml>=0.19.1",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
